requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
Flask-Caching==2.1.0
cryptography==41.0.7
Flask-Mail==0.10.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
psycopg2-binary==2.9.11
APScheduler==3.10.4